requirements.txt
fastapi
uvicorn
pydantic
mysql-connector-python
openai
PyJWT
python-dotenv